requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
Pillow>=10.0.0
typer>=0.9.0
rich>=13.0.0
python-docx>=1.1.0
